pyproject.toml
[tool.poetry]
name = "gh-linker"
version = "0.1.0"
description = ""
authors = ["laundmo <laurinschmidt2001@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
"discord.py" = { url = "https://github.com/Rapptz/discord.py/archive/45d498c1b76deaf3b394d17ccf56112fa691d160.zip" }
python-dotenv = "^0.20.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
